Add files via upload
In this project we attempted to estimate the possibility of an applicant looking for a new job in this project.
Alpha Private Limited currently has two openings vacant for the roles of JR. Data analyst
and senior data scientist. The job descriptions for both positions are as follows:
NewJob Partner is a recruitment firm that focuses on offering effective resources to a variety
of businesses in various sectors.
Alpha Private Limited has two available opportunities for JR. Data analyst and Senior Data
Scientist. The job descriptions for both positions are as follows:
Junior Data Analyst: Experience 0-3 years
Stem Graduate, Any Postgraduate
Humanities and Business degree not required
Senior Data Scientist: Experience 10 or more years
Stem Graduate, Any Postgraduate, Any PhD
Humanities and Business degree not required
Alpha Private Limited worked with NewJob partner to find the best applicant for the newly
available position.
Statement of the Problem: Determine the likelihood of a candidate looking for a new job.